requirements.txt
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
dnspython==2.2.1
email-validator==1.1.3
et-xmlfile==1.1.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Dance==5.1.0
Flask-Login==0.6.0
flask-mongoengine==1.0.0
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
mongoengine==0.24.1
numpy==1.21.5
oauthlib==3.2.0
openai==0.16.0
openpyxl==3.0.9
pandas==1.3.5
pandas-stubs==1.2.0.56
pycparser==2.21
PyJWT==2.3.0
pymongo==4.1.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
tqdm==4.64.0
typing-extensions==4.1.1
urllib3==1.26.9
URLObject==2.4.3
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.7.0
